# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  schemas.CreateProjectRequest:
    properties:
      backend_techs:
        items:
          type: string
        type: array
      delivery_date:
        type: string
      description:
        type: string
      estimated_time:
        type: integer
      frontend_techs:
        items:
          type: string
        type: array
      level:
        type: string
      quanty_max_users:
        type: integer
      title:
        type: string
    required:
    - delivery_date
    - description
    - estimated_time
    - quanty_max_users
    - title
    type: object
...
  schemas.UserIDRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
host: localhost:5000
info:
  contact:
    email: support@hublast.com
    name: API Support
  description: Project Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: HubLast-Hub API
  version: "1.0"
paths:
  /admin/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a project by its ID (admin only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates an existing project with the provided details
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated project details
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns message and updated project
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - projects
  /admin/projects/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approves a pending project (admin only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project approved successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a project
      tags:
      - projects
  /admin/projects/{id}/users:
    post:
      consumes:
      - application/json
      description: Adds a user to a specific project (admin only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID to add
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/schemas.UserIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User added to project successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add user to project
      tags:
      - projects
...
  /api/admin/projects/{id}/users:
    delete:
      consumes:
      - application/json
      description: Removes a user from a specific project (admin only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID to add
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/schemas.UserIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User removed from project successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove user from project
      tags:
      - projects
  /api/projects/{id}/join:
    post:
      consumes:
      - application/json
      description: Allows the current user to request joining a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Join request successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Request to join a project
      tags:
      - projects
  /api/projects/{id}/users:
    get:
      consumes:
      - application/json
      description: Retrieves users of a project by project ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns project users
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get users of a specific project
      tags:
      - projects
  /api/user/projects:
    get:
      consumes:
      - application/json
      description: Retrieves all projects associated with the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Returns list of user's projects
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's projects
      tags:
      - projects
...
  /projects:
    get:
      consumes:
      - application/json
      description: Retrieves all projects, optionally filtered by status
      parameters:
      - description: Filter projects by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns list of projects
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project with the provided details
      parameters:
      - description: Project details
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Returns message and created project
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - projects
  /projects/{id}/users:
    get:
      consumes:
      - application/json
      description: Retrieves users of a project by project ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns project users information
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get users of a specific project
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"